Refactor metrics namespace
